Run checker on Code.require_file/1
Closes #9362.
