+++ lib/elixir/lib/code.ex
@@ -844,7 +844,7 @@ defmodule Code do
     :elixir_code_server.call({:acquire, file})
     loaded = :elixir_compiler.file(file, fn _, _ -> :ok end)
     :elixir_code_server.cast({:required, file})
-    loaded
+    verify_loaded(loaded)
   end
 
   @doc """
@@ -890,7 +890,7 @@ defmodule Code do
       :proceed ->
         loaded = :elixir_compiler.file(file, fn _, _ -> :ok end)
         :elixir_code_server.cast({:required, file})
-        loaded
+        verify_loaded(loaded)
     end
   end
 
@@ -1083,7 +1083,8 @@ defmodule Code do
   """
   @spec compile_string(List.Chars.t(), binary) :: [{module, binary}]
   def compile_string(string, file \\ "nofile") when is_binary(file) do
-    :elixir_compiler.string(to_charlist(string), file, fn _, _ -> :ok end)
+    loaded = :elixir_compiler.string(to_charlist(string), file, fn _, _ -> :ok end)
+    Enum.map(loaded, fn {module, _map, binary} -> {module, binary} end)
   end
 
   @doc """
@@ -1096,7 +1097,8 @@ defmodule Code do
   """
   @spec compile_quoted(Macro.t(), binary) :: [{module, binary}]
   def compile_quoted(quoted, file \\ "nofile") when is_binary(file) do
-    :elixir_compiler.quoted(quoted, file, fn _, _ -> :ok end)
+    loaded = :elixir_compiler.quoted(quoted, file, fn _, _ -> :ok end)
+    Enum.map(loaded, fn {module, _map, binary} -> {module, binary} end)
   end
 
   @doc """
@@ -1116,7 +1118,8 @@ defmodule Code do
   @doc since: "1.7.0"
   @spec compile_file(binary, nil | binary) :: [{module, binary}]
   def compile_file(file, relative_to \\ nil) when is_binary(file) do
-    :elixir_compiler.file(find_file(file, relative_to), fn _, _ -> :ok end)
+    loaded = :elixir_compiler.file(find_file(file, relative_to), fn _, _ -> :ok end)
+    verify_loaded(loaded)
   end
 
   @doc """
@@ -1342,5 +1345,11 @@ defmodule Code do
     else
       raise Code.LoadError, file: file
     end
+  end
+
+  defp verify_loaded(loaded) do
+    maps_binaries = Enum.map(loaded, fn {_module, map, binary} -> {map, binary} end)
+    {loaded, _warnings} = Module.ParallelChecker.verify(maps_binaries, [])
+    loaded
   end
 end
