+++ lib/elixir/test/elixir/fixtures/checker_warning.exs
@@ -1,0 +1,3 @@
+defmodule CodeTest.CheckerWarning do
+  def foo(x) when is_atom(x) and is_list(x), do: x
+end
